Add 'try an example' button for 5 provided examples
There is now a getExampleData function that gets the example goids
and sections file data from the server. The sections file
information is submitted for the user by a hidden input.

Minor style change for the sections in the results table.

Added 2d list handling to base_results.html so examples would
be displayed